Store .node content ALSO in the neo slot
The .node pages are storing content in the main slot.
This means the SubjectRepo does not find the content, since it looks
for it in the neo slot. Ideally the .node pages would use the
neo slot instead of the main slot. Perhaps this is possible via
some other MW hook. The current hotfix saves the content from the
main slot in the neo slot as well. Likely this has unintended
side effects. Anyway, the .node pages are a hack for data entry,
pending a proper interface for regular pages