Add JSON schema stubs
For https://github.com/ProfessionalWiki/NeoWiki/issues/15

Looks like we can perhaps use an extended version of JSON Schema
to define schemas in the wiki.

For instance, you'd have pages such as Schema:Employee with a JSON
Schema that defines which properties there are and what their constraints
are. We can also stuff in additional data into the schema, such
as how these values should be formatted in our UIs, or which editing
elements to use.

https://json-schema.org/

I also has a brief look at https://shex.io/, since SheX is used
by Wikidata. The SheX excosystem seems really immature though,
and we need to be able to specify additional information for our UIs.